+++ stream-chat-android-docs/src/main/java/io/getstream/chat/docs/java/AppSettings.java
@@ -8,27 +8,12 @@ public class AppSettings {
     private Context context;
 
     /**
-     * @see <a href="https://getstream.io/chat/docs/multi_region/?language=java#client-location-settings">Client Location Settings</a>
+     * @see <a href="https://getstream.io/chat/docs/android/multi_region/?language=java#client-location-settings">Client Location Settings</a>
      */
     public void clientLocationSettings() {
-        // EU West - Dublin region
-        ChatClient dublinClient = new ChatClient.Builder("{{ api_key }}", context)
-                .baseUrl("https://chat-proxy-dublin.stream-io-api.com/")
-                .build();
-
-        // US East region
-        ChatClient usEastClient = new ChatClient.Builder("{{ api_key }}", context)
-                .baseUrl("https://chat-proxy-us-east.stream-io-api.com/")
-                .build();
-
-        // Singapore region
-        ChatClient singaporeClient = new ChatClient.Builder("{{ api_key }}", context)
-                .baseUrl("https://chat-proxy-singapore.stream-io-api.com/")
-                .build();
-
-        // Sydney region
-        ChatClient sydneyClient = new ChatClient.Builder("{{ api_key }}", context)
-                .baseUrl("https://chat-proxy-sydney.stream-io-api.com/")
+        // Set the base URL
+        ChatClient client = new ChatClient.Builder("{{ api_key }}", context)
+                .baseUrl("https://chat.stream-io-api.com/")
                 .build();
     }
 }